requirements-dev.txt
# 🏛️ Kapital Bank AI Assistant - Development Dependencies
# Additional packages needed for development, testing, and code quality

# =============================================================================
# 🧪 TESTING FRAMEWORKS
# =============================================================================

# Core testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0

# HTTP testing
httpx==0.28.1
requests-mock==1.11.0

# Performance testing
locust==2.20.0

# =============================================================================
# 🔍 CODE QUALITY & LINTING
# =============================================================================

# Code formatting
black==23.12.1
isort==5.13.2

# Linting
flake8==6.1.0
pylint==3.0.3
bandit==1.7.5

# Type checking
mypy==1.8.0
types-requests==2.31.0.20240106

# Security scanning
safety==3.0.1
semgrep==1.52.0

# =============================================================================
# 📊 MONITORING & PROFILING
# =============================================================================

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0
py-spy==0.3.14

# Debugging
pdb-attach==3.2.0
icecream==2.1.3

# =============================================================================
# 📝 DOCUMENTATION
# =============================================================================

# Documentation generation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==1.25.2

# API documentation
redoc-cli==0.13.21

# =============================================================================
# 🐳 CONTAINERIZATION & DEPLOYMENT
# =============================================================================

# Docker utilities
docker-compose==1.29.2

# Kubernetes
kubernetes==29.0.0

# =============================================================================
# 🔧 DEVELOPMENT UTILITIES
# =============================================================================

# Environment management
python-dotenv==1.1.0

# Jupyter for analysis
jupyter==1.0.0
ipykernel==6.27.1

# Database tools
sqlite-web==0.6.2

# Task automation
invoke==2.2.0

# Pre-commit hooks
pre-commit==3.6.0

# =============================================================================
# 📈 DATA ANALYSIS & VISUALIZATION
# =============================================================================

# Data analysis
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# =============================================================================
# 🔬 AI/ML DEVELOPMENT TOOLS
# =============================================================================

# Model testing
transformers==4.36.2

# Prompt engineering
langchain==0.1.0

# =============================================================================
# 🌐 API DEVELOPMENT TOOLS
# =============================================================================

# API testing
tavern==2.5.1

# Mock servers
responses==0.24.1

# OpenAPI tools
openapi-spec-validator==0.7.1

# =============================================================================
# 🔄 CI/CD TOOLS
# =============================================================================

# Coverage reporting
coverage==7.3.4
codecov==2.1.13

# Build tools
build==1.0.3
twine==4.0.2

# =============================================================================
# 🐛 DEBUGGING & DEVELOPMENT SERVERS
# =============================================================================

# Development server with auto-reload
watchdog==3.0.0

# Interactive debugging
ipdb==0.13.13

# Rich terminal output
rich==13.7.0

# =============================================================================
# 📊 LOAD TESTING & BENCHMARKING
# =============================================================================

# Load testing
aiohttp-jinja2==1.6.0
aiofiles==24.1.0

# Benchmarking
pytest-benchmark==4.0.0

# =============================================================================
# 🔐 SECURITY TESTING
# =============================================================================

# Security testing
sqlmap==1.7.12

# Dependency checking
pip-audit==2.6.2

# =============================================================================
# 📱 FRONTEND DEVELOPMENT (if needed)
# =============================================================================

# Node.js package management (if using npm tools)
nodeenv==1.8.0

# =============================================================================
# 🎯 SPECIFIC DEVELOPMENT TOOLS
# =============================================================================

# FastAPI development tools
fastapi-cli==0.0.1

# Async development
aioresponses==0.7.4

# Database migrations (if using Alembic)
alembic==1.13.1

# Redis development
fakeredis==2.20.1

# =============================================================================
# 📱 MOBILE & PWA DEVELOPMENT
# =============================================================================

# PWA manifest validation
pwa-validator==1.0.0

# =============================================================================
# 🔧 PROJECT-SPECIFIC DEVELOPMENT TOOLS
# =============================================================================

# MCP development
mcp-dev==1.0.0

# Banking API mocking
bank-api-mock==1.0.0

# Currency API testing
currency-test-utils==1.0.0

# =============================================================================
# 📊 ANALYTICS & REPORTING
# =============================================================================

# Report generation
jinja2==3.1.4
weasyprint==60.2

# Data export
openpyxl==3.1.2

# =============================================================================
# 🎨 UI/UX DEVELOPMENT TOOLS
# =============================================================================

# CSS processing (if needed)
rcssmin==1.1.1

# JavaScript minification (if needed)
rjsmin==1.2.0

# =============================================================================
# 🌍 INTERNATIONALIZATION
# =============================================================================

# Translation management
babel==2.14.0
flask-babel==4.0.0

# =============================================================================
# 📝 NOTES FOR DEVELOPERS
# =============================================================================

# Installation:
#   pip install -r requirements-dev.txt
#
# Key tools:
#   - pytest: Run tests with `pytest`
#   - black: Format code with `black .`
#   - flake8: Lint code with `flake8 .`
#   - mypy: Type check with `mypy .`
#   - locust: Load test with `locust`
#   - jupyter: Data analysis with `jupyter notebook`
#
# Pre-commit setup:
#   pre-commit install
#   pre-commit run --all-files
#
# Testing commands:
#   pytest                              # Run all tests
#   pytest --cov=.                      # Run with coverage
#   pytest -x                           # Stop on first failure
#   pytest -v                           # Verbose output
#   pytest tests/test_specific.py       # Run specific test file
#
# Code quality:
#   black . && isort . && flake8 .      # Format and lint
#   mypy . --ignore-missing-imports     # Type checking
#   bandit -r .                         # Security scanning
#   safety check                        # Dependency security
#
# Documentation:
#   sphinx-build -b html docs docs/_build  # Build docs
#   redoc-cli build openapi.json            # API docs
#
# Performance:
#   py-spy top --pid <PID>              # Profile running app
#   memory_profiler script.py           # Memory profiling
#   locust -f locustfile.py             # Load testing